Relocate m256wh VIA logic
